ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.20
langchain-openai==0.1.1
langchain-core==0.1.52
pydantic==2.10.6
pydantic-core==2.27.2
python-multipart==0.0.6
python-dotenv==1.0.0
